+++ lib/glific/templates/session_templates.ex
@@ -173,16 +173,4 @@ defmodule Glific.Templates.SessionTemplate do
   def to_minimal_map(sessiontemplate) do
     Map.take(sessiontemplate, [:id | @required_fields ++ @optional_fields])
   end
-
-  @doc """
-  List of available categories provided by whatsapp
-  """
-  @spec list_whatsapp_hsm_categories() :: [String.t()]
-  def list_whatsapp_hsm_categories do
-    [
-      "UTILITY",
-      "MARKETING",
-      "AUTHENTICATION"
-    ]
-  end
 end
