+++ lib/glific_web/schema/session_template_type.ex
@@ -6,8 +6,8 @@ defmodule GlificWeb.Schema.SessionTemplateTypes do
   use Absinthe.Schema.Notation
   import Absinthe.Resolution.Helpers, only: [dataloader: 1]
 
+  alias Glific.Templates
   alias Glific.Repo
-  alias Glific.Templates.SessionTemplate
   alias GlificWeb.Resolvers
   alias GlificWeb.Schema.Middleware.Authorize
 
@@ -142,7 +142,7 @@ defmodule GlificWeb.Schema.SessionTemplateTypes do
       middleware(Authorize, :manager)
 
       resolve(fn _, _, _ ->
-        {:ok, SessionTemplate.list_whatsapp_hsm_categories()}
+        {:ok, Templates.list_whatsapp_hsm_categories()}
       end)
     end
